# samples/server/petstore/springboot-delegate-j8/src/main/java/org/openapitools/api/UserApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.CreateUsersWithArrayInput201Response;
import org.openapitools.model.Param0Def;
import org.openapitools.model.Param1Def;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Api(value = "user", description = "the user API")
public interface UserApi {

    default UserApiDelegate getDelegate() {
        return new UserApiDelegate() {};
    }

    /**
     * POST /user/createWithArray : Creates list of users with given input array
     * 
     *
     * @param param0  (optional)
     * @param param1  (optional)
     * @return bar baz (status code 200)
     *         or foo bar (status code 201)
     */
    @ApiOperation(
        tags = { "user" },
        value = "Creates list of users with given input array",
        nickname = "createUsersWithArrayInput",
        notes = "",
        response = Param1Def.class,
        responseContainer = "List"
    )
    @ApiResponses({
        @ApiResponse(code = 200, message = "bar baz", response = Param1Def.class, responseContainer = "List"),
        @ApiResponse(code = 201, message = "foo bar", response = CreateUsersWithArrayInput201Response.class)
    })
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/user/createWithArray",
        produces = { "application/json" }
    )
    default ResponseEntity<List<Param1Def>> createUsersWithArrayInput(
        @ApiParam(value = "") @RequestHeader(value = "param0", required = false) Param0Def param0,
        @ApiParam(value = "") @RequestHeader(value = "param1", required = false) List<@Valid Param1Def> param1
    ) {
        return getDelegate().createUsersWithArrayInput(param0, param1);
    }

}