# samples/server/petstore/springboot-useoptional/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: dummy
  title: OpenAPI Petstore
  version: 1.0.0
servers:
- url: /
paths:
  /user/createWithArray:
    post:
      description: ""
      operationId: createUsersWithArrayInput
      parameters:
      - explode: false
        in: header
        name: param0
        required: false
        schema:
          $ref: '#/components/schemas/Param0Def'
        style: simple
      - explode: false
        in: header
        name: param1
        required: false
        schema:
          items:
            $ref: '#/components/schemas/Param1Def'
          type: array
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Param1Def'
                type: array
          description: bar baz
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createUsersWithArrayInput_201_response'
          description: foo bar
      summary: Creates list of users with given input array
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
components:
  responses:
    ResponseDef:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/createUsersWithArrayInput_201_response'
      description: foo bar
  schemas:
    Param0Def:
      properties:
        param0Field:
          items:
            $ref: '#/components/schemas/Param1Def'
          type: array
      type: object
    Param1Def:
      example:
        myField:
        - 0
        - 0
      properties:
        myField:
          items:
            type: integer
          type: array
      type: object
    MyIntArray:
      items:
        type: integer
      type: array
    createUsersWithArrayInput_201_response:
      properties:
        myField:
          items:
            type: integer
          type: array
      type: object